requirements.txt
wavelink==1.3.2
discord.py==2.0.1
python-dotenv==0.20.0
d20==1.1.2
